+++ R/client.R
@@ -9,18 +9,31 @@
 #' @param team_name,team_id For `get_team`, either the name or ID of the team to retrieve. Supply exactly one of these.
 #' @param shared_mbox_id,shared_mbox_name,shared_mbox_email For `get_business_outlook`, an ID/principal name/email address. Supply exactly one of these to retrieve a shared mailbox. If all are NULL (the default), retrieve your own mailbox.
 #' @param chat_id For `get_chat`, the ID of a group, one-on-one or meeting chat in Teams.
+#' @param token An AAD OAuth token object, of class `AzureAuth::AzureToken`. If supplied, the `tenant`, `app`, `scopes` and `...` arguments will be ignored. See "Authenticating with a token" below.
 #' @param ... Optional arguments that will ultimately be passed to [`AzureAuth::get_azure_token`].
 #' @details
 #' These functions provide easy access to the various collaboration services that are part of Microsoft 365. On first use, they will call your web browser to authenticate with Azure Active Directory, in a similar manner to other web apps. You will get a dialog box asking for permission to access your information. You only have to authenticate once; your credentials will be saved and reloaded in subsequent sessions.
 #'
 #' When authenticating, you can pass optional arguments in `...` which will ultimately be received by `AzureAuth::get_azure_token`. In particular, if your machine doesn't have a web browser available to authenticate with (for example if you are in a remote RStudio Server session), pass `auth_type="device_code"` which is intended for such scenarios.
 #'
-#' @section Authenticating to Microsoft 365 Business services:
-#' Authenticating to Microsoft 365 Business services (Teams, SharePoint and OneDrive for Business) has some specific complexities.
+#' ## Authenticating to Microsoft 365 Business services
+#' Authenticating to Microsoft 365 Business services (Teams, SharePoint and business OneDrive/Outlook) has some specific complexities.
 #'
 #' The default "common" tenant for `get_team`, `get_business_onedrive` and `get_sharepoint_site` attempts to detect your actual tenant from your saved credentials in your browser. This may not always succeed, for example if you have a personal account that is also a guest account in a tenant. In this case, supply the actual tenant name, either in the `tenant` argument or in the `CLIMICROSOFT365_TENANT` environment variable. The latter allows sharing authentication details with the [CLI for Microsoft 365](https://pnp.github.io/cli-microsoft365/).
 #'
-#' The default when authenticating to these services is for Microsoft365R to use its own internal app ID. As an alternative, you (or your admin) can create your own app registration in Azure: it should have a native redirect URI of `http://localhost:1410`, and the "public client" option should be enabled if you want to use the device code authentication flow. You can supply your app ID either via the `app` argument, or in the environment variable `CLIMICROSOFT365_AADAPPID`.
+#' The default when authenticating to these services is for Microsoft365R to use its own internal app ID. As an alternative, you (or your admin) can create your own app registration in Azure: for use in a local session, it should have a native redirect URI of `http://localhost:1410`, and the "public client" option should be enabled if you want to use the device code authentication flow. You can supply your app ID either via the `app` argument, or in the environment variable `CLIMICROSOFT365_AADAPPID`.
+#'
+#' ## Authenticating with a token
+#' In some circumstances, it may be desirable to carry out authentication/authorization as a separate step prior to  making requests to the Microsoft 365 REST API. This holds in a Shiny app, for example, since only the UI part can talk to the browser while the server part does the rest of the work. Another scenario is if the refresh token lifetime set by your org is too short, so that the token expires in between R sessions.
+#'
+#' In this case, you can authenticate by obtaining a new token with `AzureAuth::get_azure_token`, and passing the token object to the client function. Note that the token is accepted as-is; no checks are performed that it has the correct permissions for the service you're using.
+#'
+#' When calling `get_azure_token`, the scopes you should use are those given in the `scopes` argument for each client function, and the API host is `https://graph.microsoft.com/`. The Microsoft365R internal app ID is `d44a05d5-c6a5-4bbb-82d2-443123722380`, while that for the CLI for Microsoft 365 is `31359c7f-bd7e-475c-86db-fdb8c937548e`. However, these app IDs **only** work for a local R session; you must create your own app registration if you want to use the package inside a Shiny app.
+#'
+#' See the examples below, and also the vignette "Using Microsoft365R in a Shiny app" for a more detailed rundown on combining Microsoft365R and Shiny.
+#'
+#' ## Clearing the cache
+#' Deleting your cached credentials is a way of rebooting the authentication process, if you are repeatedly encountering errors. To do this, call [`AzureAuth::clean_token_directory`], then try logging in again. You may also need to clear your browser's cookies, if you are authenticating interactively.
 #'
 #' @return
 #' For `get_personal_onedrive` and `get_business_onedrive`, an R6 object of class `ms_drive`.
@@ -33,7 +46,7 @@
 #'
 #' [`add_methods`] for the associated methods that this package adds to the base AzureGraph classes.
 #'
-#' The "Authentication" vignette has more details on the authentication process, including troubleshooting and fixes for common problems.
+#' The "Authentication" vignette has more details on the authentication process, including troubleshooting and fixes for common problems. The "Using Microsoft365R in a Shiny app" vignette has further Shiny-specific information, including how to configure the necessary app registration in Azure Active Directory.
 #'
 #' [CLI for Microsoft 365](https://pnp.github.io/cli-microsoft365/) -- a commandline tool for managing Microsoft 365
 #' @examples
@@ -69,14 +82,25 @@
 #' get_sharepoint_site("My site")
 #' get_team("My team")
 #'
+#' # authenticating separately to working with the MS365 API
+#' scopes <- c(
+#'     "https://graph.microsoft.com/Files.ReadWrite.All",
+#'     "https://graph.microsoft.com/User.Read",
+#'     "openid", "offline_access"
+#' )
+#' app <- "d44a05d5-c6a5-4bbb-82d2-443123722380" # for local use only
+#' token <- AzureAuth::get_azure_token(scopes, "mycompany", app, version=2)
+#' get_business_onedrive(token=token)
+#'
 #' }
 #' @rdname client
 #' @export
 get_personal_onedrive <- function(app=.microsoft365r_app_id,
                                   scopes=c("Files.ReadWrite.All", "User.Read"),
+                                  token=NULL,
                                   ...)
 {
-    do_login("consumers", app, scopes, ...)$get_user()$get_drive()
+    do_login("consumers", app, scopes, token, ...)$get_user()$get_drive()
 }
 
 #' @rdname client
@@ -84,11 +108,12 @@ get_personal_onedrive <- function(app=.microsoft365r_app_id,
 get_business_onedrive <- function(tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "common"),
                                   app=Sys.getenv("CLIMICROSOFT365_AADAPPID"),
                                   scopes=c("Files.ReadWrite.All", "User.Read"),
+                                  token=NULL,
                                   ...)
 {
     app <- choose_app(app)
     scopes <- set_default_scopes(scopes, app)
-    do_login(tenant, app, scopes, ...)$get_user()$get_drive()
+    do_login(tenant, app, scopes, token, ...)$get_user()$get_drive()
 }
 
 #' @rdname client
@@ -98,12 +123,13 @@ get_sharepoint_site <- function(site_name=NULL, site_url=NULL, site_id=NULL,
                                 app=Sys.getenv("CLIMICROSOFT365_AADAPPID"),
                                 scopes=c("Group.ReadWrite.All", "Directory.Read.All",
                                          "Sites.ReadWrite.All", "Sites.Manage.All"),
+                                token=NULL,
                                 ...)
 {
     assert_one_arg(site_name, site_url, site_id, msg="Supply exactly one of site name, URL or ID")
     app <- choose_app(app)
     scopes <- set_default_scopes(scopes, app)
-    login <- do_login(tenant, app, scopes, ...)
+    login <- do_login(tenant, app, scopes, token, ...)
 
     if(!is.null(site_name))
     {
@@ -124,11 +150,12 @@ list_sharepoint_sites <- function(tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "c
                                   app=Sys.getenv("CLIMICROSOFT365_AADAPPID"),
                                   scopes=c("Group.ReadWrite.All", "Directory.Read.All",
                                            "Sites.ReadWrite.All", "Sites.Manage.All"),
+                                  token=NULL,
                                   ...)
 {
     app <- choose_app(app)
     scopes <- set_default_scopes(scopes, app)
-    login <- do_login(tenant, app, scopes, ...)
+    login <- do_login(tenant, app, scopes, token, ...)
 
     login$get_user()$list_sharepoint_sites()
 }
@@ -139,12 +166,13 @@ get_team <- function(team_name=NULL, team_id=NULL,
                      tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "common"),
                      app=Sys.getenv("CLIMICROSOFT365_AADAPPID"),
                      scopes=c("Group.ReadWrite.All", "Directory.Read.All"),
+                     token=NULL,
                      ...)
 {
     assert_one_arg(team_name, team_id, msg="Supply exactly one of team name or ID")
     app <- choose_app(app)
     scopes <- set_default_scopes(scopes, app)
-    login <- do_login(tenant, app, scopes, ...)
+    login <- do_login(tenant, app, scopes, token, ...)
 
     if(!is.null(team_name))
     {
@@ -166,11 +194,12 @@ get_team <- function(team_name=NULL, team_id=NULL,
 list_teams <- function(tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "common"),
                        app=Sys.getenv("CLIMICROSOFT365_AADAPPID"),
                        scopes=c("Group.ReadWrite.All", "Directory.Read.All"),
+                       token=NULL,
                        ...)
 {
     app <- choose_app(app)
     scopes <- set_default_scopes(scopes, app)
-    login <- do_login(tenant, app, scopes, ...)
+    login <- do_login(tenant, app, scopes, token, ...)
 
     login$get_user()$list_teams()
 }
@@ -179,9 +208,10 @@ list_teams <- function(tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "common"),
 #' @export
 get_personal_outlook <- function(app=.microsoft365r_app_id,
                                  scopes=c("Mail.Send", "Mail.ReadWrite", "User.Read"),
+                                 token=NULL,
                                  ...)
 {
-    do_login("consumers", app, scopes, ...)$get_user()$get_outlook()
+    do_login("consumers", app, scopes, token, ...)$get_user()$get_outlook()
 }
 
 #' @rdname client
@@ -190,12 +220,13 @@ get_business_outlook <- function(tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "co
                                  app=.microsoft365r_app_id,
                                  shared_mbox_id=NULL, shared_mbox_name=NULL, shared_mbox_email=NULL,
                                  scopes=c("User.Read", "Mail.Send", "Mail.ReadWrite"),
+                                 token=NULL,
                                  ...)
 {
     if(!is.null(shared_mbox_id) || !is.null(shared_mbox_name) || !is.null(shared_mbox_email))
         scopes <- c(scopes, "Mail.Send.Shared", "Mail.ReadWrite.Shared")
 
-    do_login(tenant, app, scopes, ...)$
+    do_login(tenant, app, scopes, token, ...)$
         get_user(user_id=shared_mbox_id, name=shared_mbox_name, email=shared_mbox_email)$
         get_outlook()
 }
@@ -206,9 +237,10 @@ get_chat <- function(chat_id,
                      tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "common"),
                      app=.microsoft365r_app_id,
                      scopes=c("User.Read", "Directory.Read.All", "Chat.ReadWrite"),
+                     token=NULL,
                      ...)
 {
-    do_login(tenant, app, scopes, ...)$get_user()$get_chat(chat_id)
+    do_login(tenant, app, scopes, token, ...)$get_user()$get_chat(chat_id)
 }
 
 
@@ -217,30 +249,26 @@ get_chat <- function(chat_id,
 list_chats <- function(tenant=Sys.getenv("CLIMICROSOFT365_TENANT", "common"),
                        app=.microsoft365r_app_id,
                        scopes=c("User.Read", "Directory.Read.All", "Chat.ReadWrite"),
+                       token=NULL,
                        ...)
 {
-    do_login(tenant, app, scopes, ...)$get_user()$list_chats()
+    do_login(tenant, app, scopes, token, ...)$get_user()$list_chats()
 }
 
 
-.ms365_login_env <- new.env()
-
-do_login <- function(tenant, app, scopes, ...)
+do_login <- function(tenant, app, scopes, token, ...)
 {
-    hash <- function(...)
+    # bypass AzureGraph login caching if token provided
+    if(!is.null(token))
     {
-        as.character(openssl::md5(serialize(list(...), NULL)))
+        if(!AzureAuth::is_azure_token(token))
+            stop("Invalid token object supplied")
+        return(ms_graph$new(token=token))
     }
 
-    login_id <- hash(tenant, app, scopes, ...)
-    login <- .ms365_login_env[[login_id]]
-    if(is.null(login) || !inherits(login, "ms_graph"))
-    {
-        login <- try(get_graph_login(tenant, app=app, scopes=scopes, refresh=FALSE), silent=TRUE)
-        if(inherits(login, "try-error"))
-            login <- create_graph_login(tenant, app=app, scopes=scopes, ...)
-        .ms365_login_env[[login_id]] <- login
-    }
+    login <- try(get_graph_login(tenant, app=app, scopes=scopes, refresh=FALSE), silent=TRUE)
+    if(inherits(login, "try-error"))
+        login <- create_graph_login(tenant, app=app, scopes=scopes, ...)
     login
 }